requirements.txt
# Core Dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
alembic>=1.12.1
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
redis>=5.0.1
celery>=5.3.4

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1
cryptography>=41.0.7

# Data Processing
pandas>=2.1.3
numpy>=1.25.2
PyPDF2>=3.0.1
openpyxl>=3.1.2
xlsxwriter>=3.1.9
python-multipart>=0.0.6

# Financial Calculations
scipy>=1.11.4
pyxirr>=0.10.4

# HTTP Clients & External APIs
httpx>=0.25.2
requests>=2.31.0

# Configuration & Environment
pydantic>=2.5.0
pydantic-settings>=2.0.3
python-dotenv>=1.0.0

# Logging & Monitoring
structlog>=23.2.0
sentry-sdk>=1.38.0

# Validation & Utilities
email-validator>=2.1.0
python-slugify>=8.0.1
python-dateutil>=2.8.2
pytz>=2023.3

# Testing (Development)
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
httpx>=0.25.2
faker>=20.1.0

# Code Quality (Development)
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.1
pre-commit>=3.5.0

# Documentation
mkdocs>=1.5.3
mkdocs-material>=9.4.8